Move config parser definition from datastore to own module
