+++ OTAnalytics/application/datastore.py
@@ -4,6 +4,7 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Iterable, Optional, Sequence, Tuple
 
+from OTAnalytics.application.parser.config_parser import ConfigParser
 from OTAnalytics.application.project import Project
 from OTAnalytics.application.use_cases.export_events import EventListExporter
 from OTAnalytics.domain.event import Event, EventRepository
@@ -207,38 +208,6 @@ class TrackVideoParser(ABC):
         Returns:
             Tuple[list[TrackId], list[Video]]: track ids and the corresponding videos
         """
-        pass
-
-
-@dataclass(frozen=True)
-class OtConfig:
-    project: Project
-    videos: Sequence[Video]
-    sections: Sequence[Section]
-    flows: Sequence[Flow]
-
-
-class ConfigParser(ABC):
-    """
-    Serialize and parse config files generated by OTConfig
-    """
-
-    @abstractmethod
-    def parse(
-        self,
-        file: Path,
-    ) -> OtConfig:
-        pass
-
-    @abstractmethod
-    def serialize(
-        self,
-        project: Project,
-        video_files: Iterable[Video],
-        sections: Iterable[Section],
-        flows: Iterable[Flow],
-        file: Path,
-    ) -> None:
         pass
