+++ tests/OTAnalytics/application/test_datastore.py
@@ -8,17 +8,16 @@ from numpy import array, int32
 from PIL import Image
 
 from OTAnalytics.application.datastore import (
-    ConfigParser,
     Datastore,
     EventListParser,
     FlowParser,
-    OtConfig,
     TrackParser,
     TrackToVideoRepository,
     TrackVideoParser,
     VideoMetadata,
     VideoParser,
 )
+from OTAnalytics.application.parser.config_parser import ConfigParser, OtConfig
 from OTAnalytics.application.project import Project
 from OTAnalytics.domain.event import EventRepository
 from OTAnalytics.domain.flow import Flow, FlowRepository
